+++ kernel/src/mm/mod.rs
@@ -22,6 +22,7 @@ use self::{
 pub mod allocator;
 pub mod c_adapter;
 pub mod early_ioremap;
+pub mod fault;
 pub mod init;
 pub mod kernel_mapper;
 pub mod madvise;
